Refactors server code, fixes issues with testing environment loading in sandbox.
Refactors sandbox server script to make it easier to test and debug.

Testing index page is now loaded with the entry file from the sandbox.

isAdmin will query the :3000 node server for a user, then check it's isAdmin flag and return it.
isADmin should be invoked to check the sandbox bin at the beginning of update sandbox.
There is an issue with this right now.